@@ -2,8 +2,4 @@
   <div></div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from "vue";
-
-export default defineComponent({});
-</script>
+<script setup lang="ts"></script>
