@@ -2,16 +2,12 @@ function formatQuery(queryParams?: Record<string, string | number>) {
   return (
     (queryParams &&
       Object.keys(queryParams).length &&
-      "?" +
+      '?' +
         Object.keys(queryParams)
-          .map((key) =>
-            queryParams[key]
-              ? `${key}=${encodeURIComponent(queryParams[key])}`
-              : key
-          )
-          .join("&")) ||
-    ""
-  );
+          .map((key) => (queryParams[key] ? `${key}=${encodeURIComponent(queryParams[key])}` : key))
+          .join('&')) ||
+    ''
+  )
 }
 
 export async function post(
@@ -19,15 +15,15 @@ export async function post(
   data: Record<string, string | number>,
   queryParams?: Record<string, string | number>
 ) {
-  const query = formatQuery(queryParams);
+  const query = formatQuery(queryParams)
 
   const response = await fetch(`${url}${query}`, {
     body: JSON.stringify(data),
     headers: {
-      "Content-Type": "application/json",
+      'Content-Type': 'application/json'
     },
-    method: "POST",
-  });
+    method: 'POST'
+  })
 
-  return await response?.json();
+  return await response?.json()
 }
