@@ -1,46 +1,43 @@
-import { reactive } from "vue";
+import { reactive } from 'vue'
 
 export const recipeStore: any = reactive({
   recipe: {
-    name: "Pilzpfanne mit Hirtenkäse",
+    name: 'Pilzpfanne mit Hirtenkäse',
     ingredients: [
       {
-        amount: "200g",
-        description: "Speisequark",
+        amount: '200g',
+        description: 'Speisequark'
       },
       {
-        amount: "1 Packung",
-        description: "Hirtenkäse",
+        amount: '1 Packung',
+        description: 'Hirtenkäse'
       },
       {
-        amount: "400g",
-        description: "Champignons",
+        amount: '400g',
+        description: 'Champignons'
       },
       {
-        amount: "1 Becher",
-        description: "Crème fraîche",
-      },
+        amount: '1 Becher',
+        description: 'Crème fraîche'
+      }
     ],
     steps: [
       {
-        description: "Champignons putzen und in Scheiben schneiden.",
+        description: 'Champignons putzen und in Scheiben schneiden.'
       },
       {
-        description: "Die Champignons in einer Pfanne mit etwas Öl anbraten.",
+        description: 'Die Champignons in einer Pfanne mit etwas Öl anbraten.'
       },
       {
-        description:
-          "Quark und Crème fraîche in die Pfanne geben und verrühren.",
+        description: 'Quark und Crème fraîche in die Pfanne geben und verrühren.'
       },
       {
-        description:
-          "Den Hirtenkäse über die Pilzmasse bröseln und schmelzen lassen.",
+        description: 'Den Hirtenkäse über die Pilzmasse bröseln und schmelzen lassen.'
       },
       {
-        description:
-          "Mit Salz und Pfeffer abschmecken und nach Belieben mit Kräutern verfeinern.",
-      },
-    ],
+        description: 'Mit Salz und Pfeffer abschmecken und nach Belieben mit Kräutern verfeinern.'
+      }
+    ]
   },
-  imageUrl: "https://via.placeholder.com/150",
-});
+  imageUrl: 'https://via.placeholder.com/150'
+})
