@@ -1,3 +1,3 @@
 export const SUGGEST_RECIPE_URL =
-  process.env.VUE_APP_SUGGEST_RECIPE_IMAGE_URL ||
-  `${location.protocol}//${window.location.hostname}:3000/suggest-recipe`;
+  import.meta.env.VITE_SUGGEST_RECIPE_IMAGE_URL ||
+  `${location.protocol}//${window.location.hostname}:3000/suggest-recipe`
